requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-Login==0.6.3
Flask-Babel==4.0.0
SQLAlchemy==2.0.34
alembic==1.13.2
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
Pillow==10.4.0
python-dotenv==1.0.1
gunicorn==22.0.0
pytest==8.3.3
